feat(tests): examples -> workloads
Closes RNDSTROPPY-51
