+++ internal/runner/script_extractor.go
@@ -1,4 +1,5 @@
 // Package runner provides functionality to run TypeScript benchmark scripts with k6.
+// TODO: extractor works if helpers (as parse_sql_2) with '.ts' extension, but k6 needs '.js'. it should be consistant.
 package runner
 
 import (
