+++ workloads/execute_sql/execute_sql.ts
@@ -162,36 +162,36 @@ lookup(workloads, "workload", "transaction", "update_and_log").params.push(
 // Apply generator ranges from SQL syntax
 apply_generators_ranges(workloads);
 
+declare const __ENV: Record<string, string | undefined>;
+
 // Initialize driver with GlobalConfig
 // This is called at the top level to configure the driver
 stroppy.defineConfig(
-  GlobalConfig.toBinary(
-    GlobalConfig.create({
-      driver: {
-        url: __ENV.DRIVER_URL || "postgres://postgres:postgres@localhost:5432",
-        driverType: 1,
-        dbSpecific: {
-          fields: [
-            {
-              type: { oneofKind: "string", string: "error" },
-              key: "trace_log_level",
-            },
-            {
-              type: { oneofKind: "string", string: "5m" },
-              key: "max_conn_lifetime",
-            },
-            {
-              type: { oneofKind: "string", string: "2m" },
-              key: "max_conn_idle_time",
-            },
-            { type: { oneofKind: "int32", int32: 1 }, key: "max_conns" },
-            { type: { oneofKind: "int32", int32: 1 }, key: "min_conns" },
-            { type: { oneofKind: "int32", int32: 1 }, key: "min_idle_conns" },
-          ],
-        },
+  GlobalConfig.create({
+    driver: {
+      url: __ENV.DRIVER_URL || "postgres://postgres:postgres@localhost:5432",
+      driverType: 1,
+      dbSpecific: {
+        fields: [
+          {
+            type: { oneofKind: "string", string: "error" },
+            key: "trace_log_level",
+          },
+          {
+            type: { oneofKind: "string", string: "5m" },
+            key: "max_conn_lifetime",
+          },
+          {
+            type: { oneofKind: "string", string: "2m" },
+            key: "max_conn_idle_time",
+          },
+          { type: { oneofKind: "int32", int32: 1 }, key: "max_conns" },
+          { type: { oneofKind: "int32", int32: 1 }, key: "min_conns" },
+          { type: { oneofKind: "int32", int32: 1 }, key: "min_idle_conns" },
+        ],
       },
-    }),
-  ),
+    },
+  }),
 );
 
 // Setup function: create schema and load data
