+++ workloads/tpcds/tpcds-scale-10.sql
