+++ workloads/tpcds/tpcds-scale-10000.sql
