+++ workloads/tpcds/tpcds-scale-300.sql
