+++ workloads/tpcds/tpcds-scale-3000.sql
