+++ workloads/tpcds/tpcds-scale-50000.sql
